Added error and logging interceptors

+++ src/idea/idea.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { IdeaEntity } from './idea.entity';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -16,9 +16,15 @@ export class IdeaService {
   }
 
   async read(id: string) {
-    return await this.ideaRepository.findOne({
+    const idea = await this.ideaRepository.findOne({
       where: { id },
     });
+
+    if (!idea) {
+      throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+    }
+
+    return idea;
   }
 
   async create(data: IdeaDto) {
@@ -30,14 +36,30 @@ export class IdeaService {
   }
 
   async update(id: string, data: Partial<IdeaDto>) {
+    const idea = await this.ideaRepository.findOne({
+      where: { id },
+    });
+
+    if (!idea) {
+      throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+    }
+
     await this.ideaRepository.update({ id }, data);
 
     return await this.ideaRepository.findOne({ id });
   }
 
   async destroy(id: string) {
+    const idea = await this.ideaRepository.findOne({
+      where: { id },
+    });
+
+    if (!idea) {
+      throw new HttpException('Not found', HttpStatus.NOT_FOUND);
+    }
+
     await this.ideaRepository.delete({ id });
 
-    return { deleted: true };
+    return idea;
   }
 }
